Update singularity hub badge
hey @stigrj ! Early this week when Singularity Hub converts to an archive, the previously existing image for the badge will no longer work, so I've updated it to live on the singularityhub-docs (which I'm keeping for posterity). I want to make sure your image does not 404 so I'm doing it before that.